add http_proto and grpcclient
